Moved batch processing code to new file batchmodel.py. Changed functionality of BatchFile_from_DataFrame() so that the index of the dataframe is NOT reset by default.
